move configValueConfigDescriptor into package object, denest ConfigValuePropertyType, fix zio.config.pureconfig package directory
