+++ pureconfig/shared/src/test/scala/zio/config/pureconfig/PureconfigTest.scala
@@ -1,13 +1,13 @@
 package zio.config.pureconfig
 
-import zio.config._
-import zio.test.{Gen, checkM}
 import zio.ZIO
+import zio.config._
 import zio.test.Assertion._
-import zio.test._
+import zio.test.{Gen, checkM, _}
+import zio.random.Random
 
 object PureconfigTest extends BaseSpec {
-  override val spec =
+  override val spec: Spec[TestConfig with Random with Sized, TestFailure[Serializable], TestSuccess] =
     suite("Pureconfig support")(
       testM("vector roundtrip") {
         checkM(Gen.vectorOf(Gen.anyString)) { v =>
@@ -16,8 +16,8 @@ object PureconfigTest extends BaseSpec {
             for {
               written <- ZIO.fromEither(write(cfg, v))
               reread  <- ZIO.fromEither(
-                read(cfg from ConfigSource.fromPropertyTree(written, "tree", LeafForSequence.Valid))
-              )
+                           read(cfg from ConfigSource.fromPropertyTree(written, "tree", LeafForSequence.Valid))
+                         )
             } yield reread
 
           assertM(v2)(equalTo(v))
